fix: remove main, and move actor to main
