--- actor/go.mod
@@ -1,13 +1,0 @@
-module github.com/taiyoh/sqsd/actor
-
-go 1.16
-
-require (
-	github.com/AsynkronIT/protoactor-go v0.0.0-20210630041229-7432bbe403da
-	github.com/aws/aws-sdk-go v1.40.2
-	github.com/caarlos0/env/v6 v6.6.2
-	github.com/golang/protobuf v1.5.2
-	github.com/stretchr/testify v1.7.0
-	google.golang.org/grpc v1.39.0
-	google.golang.org/protobuf v1.27.1
-)
