+++ aws_setup_test.go
