+++ go.mod
@@ -2,16 +2,13 @@ module github.com/taiyoh/sqsd
 
 go 1.16
 
-replace github.com/taiyoh/sqsd/actor => ./actor
-
 require (
 	github.com/AsynkronIT/protoactor-go v0.0.0-20210630041229-7432bbe403da
 	github.com/aws/aws-sdk-go v1.40.2
 	github.com/caarlos0/env/v6 v6.6.2
-	github.com/fukata/golang-stats-api-handler v1.0.0
-	github.com/hashicorp/logutils v1.0.0
+	github.com/golang/protobuf v1.5.2
 	github.com/joho/godotenv v1.3.0
-	github.com/kelseyhightower/envconfig v1.4.0
-	github.com/taiyoh/sqsd/actor v0.0.0-00010101000000-000000000000
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	github.com/stretchr/testify v1.7.0
+	google.golang.org/grpc v1.39.0
+	google.golang.org/protobuf v1.27.1
 )
