+++ locker_test.go
