--- message_consumer_export_test.go
@@ -1,5 +1,0 @@
-package sqsd
-
-func (c *MessageConsumer) ChangeInvoker(invoker WorkerInvoker) {
-	c.invoker = invoker
-}
