--- queue_tracker_export_test.go
@@ -1,9 +1,0 @@
-package sqsd
-
-func (t *QueueTracker) Find(id string) (Queue, bool) {
-	data, exists := t.currentWorkings.Load(id)
-	if !exists {
-		return Queue{}, exists
-	}
-	return data.(Queue), exists
-}
