--- version.go
@@ -1,8 +1,0 @@
-package sqsd
-
-var version string
-
-// GetVersion returns version string embedded when build
-func GetVersion() string {
-	return version
-}
